requirements.txt
python-telegram-bot==20.8
pandas==2.2.3
openpyxl==3.1.5
python-dateutil==2.9.0
watchdog==3.0.0